app/requirements.txt
streamlit
PIL 
pandas
numpy
matplotlib.pyplot
data_processing_app
sklearn
xgboost
pickle